openmrs_chatbot/requirements.txt
mysql-connector-python==8.2.0
chromadb==0.4.14
langchain==0.0.352
ollama==0.1.9
pypdf==3.17.1
python-dotenv==1.0.0
flask==3.0.0
markdown==3.5.1
python-dateutil==2.8.2
sentence-transformers==2.2.2
numpy>=1.21.0
scikit-learn>=1.0.0
